Move the singular web page related structures out of model/
