--- singular/model/web.go
@@ -1,36 +1,0 @@
-/*
-Copyright 2016 - 2017 Huawei Technologies Co., Ltd. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package model
-
-import "html/template"
-
-type HtmlDeployment struct {
-	ID          int64
-	InfraName   string
-	InfraLogo   string
-	InfraLog    template.HTML
-	Components  []HtmlComponent
-	StatusIcon  string
-	StatusColor string
-}
-
-type HtmlComponent struct {
-	Name          string
-	Log           template.HTML
-	ImageSrc, Alt string
-	Width, Height int
-}
